Fix: libstonithd: Change stonith-event XML output.
This brings the output more in line with what's happening in
print_stonith_action in crm_mon.  The following changes have been made
to the XML output:

* The main XML node is now fence_event instead of stonith-event.  This
keeps compatibility with crm_mon's schema.

* The 'when' attribute is now named 'completed'.  This is also for
crm_mon compatibility purposes.

* The XML output now includes a new extended-status attribute.  This is
because anything that's not failed or done is "pending", and some users
may wish to have more information than that.