Fix: libcrmcommon: export logfile environment variable if using default
When not explicitly specifying a log, Pacemaker before 2.0.0 would set the
logfile environment variables when checking the corosync configuration for
logging options. Since Pacemaker 2 does not use corosync's logging options
anymore, this step was skipped. However, some resource agents relied on the
environment variables to enable logging, so this introduced a regression.

This fixes it by exporting the default log name when (successfully) used.