Feature: tools: Add curses formatting for stonith events.
This involves some code duplication with the built-in text formatter.
In fact, the way the text and curses formatters work is exactly the same
so we could theoretically just reuse the same code and pass in the
appropriate pcmk__output_t.  However, that would mean these functions
could no longer be static and I don't really want to make them more
broadly available.