Low: libstonithd: handle API creation errors better
Our general policy toward memory allocation errors so far has been to not
bother checking, and let the first use cause a segmentation fault, with the
idea that a node that's critically low on memory should get fenced as soon as
possible, not to mention the code clutter.

On the other hand, checking for allocation failure is cleaner, and can help
rule out memory issues when debugging strange problems. This commit takes
that route for stonith API creation.