Low: pacemaker-based: defer to fully customizable PAM service config
This affords us (but also the users) some wiggle space for customization
going forwards.  It deliberately points to the more global configuration
shipped with PAM stack that is in particular to deal with password-based
authentication (presumably pam_unix.so plays the key role there, so that
standard local user dataset gets used, which is exactly what we want,
since that's in-line with the other check pertaining the same "user"
at hand, that is, whether such user belongs to the "haclient" group
per the counterpart local dataset[*]).

[*] note that it woud be quite a stretch when totally disjoint sets
    (as allowed with PAM) were examined using the same user identifier
    as a key, since no overlap/correlation is guranteed then