Fix: controller: use op status, not rc, for executor disconnection
Previously, if an action were requested for an executor (local or remote) that
the controller does not have a connection to, the action's rc would be set to
PCMK_OCF_CONNECTION_DIED and its op status to PCMK_LRM_OP_ERROR.

This was undesirable for a couple reasons: PCMK_OCF_CONNECTION_DIED is a
nonstandard extension to the OCF return codes, which can confuse users
trying to look up the meaning or interpret cluster status output; and it really
is an operation execution status and not an operation result.

This changes the result to PCMK_OCF_UNKNOWN_ERROR with a new op status
PCMK_LRM_OP_NOT_CONNECTED. The new codes are mapped to the old ones for older
DCs that don't understand them.