Fix: controller: use op status, not rc, for execution in invalid state
Previously, if an action were requested while the controller cannot execute actions
(i.e. shutdown), the action's rc would be set to CRM_DIRECT_NACK_RC and its op
status to PCMK_LRM_OP_ERROR.

This was undesirable for a couple reasons: rc should only be OCF return codes,
and it really is an operation execution status and not an operation result.

This changes the result to PCMK_OCF_UNKNOWN_ERROR with a new op status
PCMK_LRM_OP_INVALID. The new codes are mapped to the old ones for older
DCs that don't understand them.