Maint: introduce mock/experimental skeletons, now for cib (based) daemon
[this is a rebase from 2.0, occurrences of "based" mean "cib", but they
are not changed everywhere, especially not in the name of the mocked
daemon, for easy comparability amongst the branches]

This is meant for multiple purposes (non-exhaustive):

- simulating scenarios that are rather hard to achieve in practice
  (overwhelming number of nodes or artificially capped resources)

- robustness testing (will other daemons survive intermittently, yet
  frequently missing responses to the based queries?  to what extent
  do they rely on some causality relations like notify-after-write?
  what if a chance of continuously changed CIB contents causing
  troubles?)

- anything not fitting the former two categories

Another option would be to start from a full-fledged daemon, cutting it
to pieces self-contained, patchable pieces, but since in case of based,
it has grown rather organically as a monolith, it would be substantially
more hassle than start grab the most vital pieces and start with a clean
room.  The aim is to keep additional functionality on top of said
skeleton, and let the additional functionality stay apart in opt-in
loosly coupled modules.  We shall see how this works out.