Log: based: make "preventive rename" on CIB access fault path-specific
Not so helpful to observe

> Archiving unusable file /var/lib/pacemaker/cib/cib.xml as
> /var/lib/pacemaker/cib/cib.auto.XXXXXX

(not the end of the world either, especially if the chronology can
be reconstructed through timestamps)

The pessimistic case of mkstemp failing is already addressed with
a follow-up message (it's not so important to distinguish the
cause, since we know statically that trivial EINVAL won't be
that case, and anything else is expected to be a fallout of
about the same circumstance, e.g. filesystem/dir perms).