Feature: libcrmcommon: Add an error message.
And use this message in stonith_admin.  For text output, this prints on stderr
immediately.  For XML output, this keeps track of all the errors that have
occurred and adds them at the end.

The XML schema has also slightly changed.  The message associated with
the return code is now an attribute on the status element, rather than
being added as the text child of that element.  This is to give it a
little more context and prevent confusion with any error messages.

In stonith_admin, use this new message and the info message to format most of
what's printed out in stonith_admin.  A couple things still don't get formatted
but that's because formatted output is not yet available.