Refactor: libcrmcommon: Remove unneeded code.
pcmk__parse_output_args is no longer necessary.  stonith_admin was the only
program using it (as it's the only program that has been converted to using
formatted output) and with glib-based argument processing, it is no longer
necessary.

Before formatted output can be added to other programs in the future, they
will first need to be converted to use glib for argument processing.  This
doesn't seem like too big a requirement as this will also need to be done
before the formatted output help can be displayed, too.