Fix: tools: Use a copy of argv in stonith_admin.
g_option_context_parse will consume items in argv as it processes them,
which is bad because we need that array for pcmk__output_new and
crm_log_args.  We also need to do command line processing before being
able to call pcmk__output_new.

Thus, just copy argv and free the copy when we exit.