Feature: libcrmcommon: Add a "none" output formatter.
All the functions in this formatter simply do nothing, generating no
output.  It has the same API as all the other formatters, which makes it
useful as a default case when no other formatter has been chosen.  It
could also be useful if exposed to users as a simple way of turning off
output from a tool.