Low: libpe_status: check for stops correctly when unpacking migration
When unpacking resource migration history, e89ac6d added a check for whether a
stop happened after the migration, skipping the rest of the unpacking if so.

However, it mistakenly searched by node ID instead of uname, so it was
completely ineffective for most cluster nodes. It would have accidentally
worked for Pacemaker Remote nodes and for cluster nodes whose ID was explicitly
configured the same as their unames (as so happened in some of our scheduler
regression tests).

Since we no longer store more than one successful history event for a resource
on a node, we really could update the tests and remove this check altogether,
but it may serve some useful purpose as a fail-safe and future-proofing.