Fix: scheduler: properly detect dangling migrations
Previously, a failed migrate_to with a stop but no migrate_from on the target
would always lead to the source being considered "dangling" (which would force
a stop on it).

However, that is incorrect if there was a later successful start or migration
back to the node. In that case, the scheduler would get into a loop of stopping
the resource, starting it at the next transition, then stopping it again at the
next transition.

Now, the source is not marked "dangling" in that case.