Feature: libcrmcommon: Add a formatted outputter for HTML.
This uses libxml to generate HTML, which results in overly verbose HTML
but keeps us from having to generate it by hand.  Basically, everything
will have to be in a div or span if it's going to be the child of
something else.

The main consumer of this code will be the output of crm_mon, though it
should also work for stonith_admin and any other tools that switch to
using formatted output.