Feature: libcrmcommon: Add params to finish function.
This patch adds two features to the finish function for output
formatters:

* A print parameter, which controls whether the function should actually
print anything.  Confusingly, this will only apply to things like XML
and HTML that build up a structure as they go along.  The text formatter
(and other future console-oriented formatters) will not have their
behavior changed by this parameter.

* A copy_dest parameter, which allows the caller to grab a copy of
whatever internal structure the formatter is building up.  For XML and
HTML, this will be the root xmlNode.  For other formatters, it will be
NULL.