Feature: tools: Use glib for cmdline handling in crm_mon.
This commit consists of two main pieces.  The first is simply converting
to using glib.  This involves adding a bunch of callbacks, building the
GOptionEntry tables, and converting the flag setting after command line
processing is complete.

The second part is changing all the flags from being standalone to being
part of the options structure.

Note that this commit adds the start of formatted output code (a call to
pcmk__register_formats), but this is only necessary for completing the
glib code.  Formatted output is currently not supported.