Feature: tools: Use glib for cmdline handling in crm_mon.
This commit consists of two main pieces.  The first is simply converting
to using glib.  This involves adding a bunch of callbacks, building the
GOptionEntry tables, and converting the flag setting after command line
processing is complete.

The second part is changing all the flags from being standalone to being
part of the options structure.

One potential change in this patch is that the various mode command line
arguments (-h, -X, -w, -s) are no longer mutually exclusive.  Instead,
whichever is given last on the command line will be the one to take
effect.

See rhbz#1699978