Feature: tools: Add initial formatted output options to crm_mon.
This adds the command line options necessary to get formatted output
started, but not much will get printed at the moment due to lack of
support in the printing functions.

Note that a lot of care needs to be taken to ensure the new command line
options and old command line options both do the same thing.