Feature: libcrmcommon:  Add two more XML-related functions.
These fill in some gaps in API coverage.  We had functions that create
nodes and text nodes, but they wanted a parent instead of just using the
current list parent.  We also had a function that used the current list
parent, but then made the new node the new parent.

We had nothing that just created nodes as children of the current list
parent and that was all.  This commit adds those.