Feature: libcrmcommon:  Add pcmk_create_html_node.
This function is similar to pcmk__output_create_html_node except it adds
to any given parent, not just to whatever is in the formatted output
structure.  It is therefore more generally useful.

For the moment, this does not have any users but it will soon as more
formatted output messages are added.