Fix: st_client: Reword stonith_event_text.
This brings the text output in line with the text output in
print_stonith_action in crm_mon.  This will allow the two to be combined
eventually.

Previously, the text output was being used by extra/resources/controld
to check if an event was pending or not.  However, that has been
switched over to use the XML output so the text output is free to
change.