Tests: cts-fencing.in: Run xmllint on stonith_admin output.
The find_validator function allows for other things besides xmllint to
be used for validating the result of each stonith_admin run.  If none is
found, validating will be skipped.

Additionally, add_cmd now supports a validate= parameter because some
tests that are added via this function should not be run through
xmllint.