Refactor: tools: Add a mon_state_t type.
This packages up the destination file pointer and mon_output_* enum into
a single structure that is then passed as an argument to all the many
print_* functions.  In the future, a pcmk__output_t can also be added to
this structure when the formatted output conversion starts happening.

The purpose of this is to contain all the data necessary for both
printing with the current output style and for printing with the
formatted output code in a single place.  This allows changing one
function at a time from the current style to the formatted output style.

When that conversion is completed, this type can go away and be replaced
with just a pcmk__output_t parameter.