Low: executor: stonith probes should fail only if previously registered
343ae2a4 made stonith probes return an error if the executor's fencer
connection was lost. However this is broader than necessary; we only need
errors to be returned for devices that were registered. Any that weren't
registered can still be assumed to be not registered.

There's a theoretical possibility that the fencer connection could somehow be
severed and some other entity register a device between then and when the
executor reconnects. But that is not a realistic scenario, whereas probing a
fence device on a node where the fencer respawned sometime in the past is.