Fix: libcrmcommon: Allow glib cmdline parsing without formatted output.
If pcmk__build_arg_context is given NULL for the fmts argument, it simply
will not add the formatted output command line options.  Care should be
taken to make sure that pcmk__register_formats is not called if NULL is
given, as calling that function will result in other formatted output
options potentially being added.  The result could be confusing for a
user.

Also, get rid of output_ty_desc.  It no longer serves any purpose and is
entirely duplicated by the description being built in
pcmk__build_arg_context.