Fix: libcrmcommon:  Restore getopt behavior in stonith_admin.
This patch introduces a new pcmk__cmdline_preproc function which
restores getopt's single character argument behavior that was lost when
switching to glib.  getopt recognizes a string like "-xval" to be the -x
command line option with a value of val.  glib would recognize that as
the four command line options -x, -v, -a, and -l.

Passing the command line through this new function converts it into
something that glib understands.  See the documentation and comments on
this function for more details.