Refactor: tools: Pass output_format as an arg where possible.
In general, it is now an argument to any print_* function since those
will eventually be moving out into their own file and I don't want to
make global variables like output_format visible across multiple files.
It remains global, however, primarily because there are several places
in crm_mon.c where it is needed but cannot be passed as an argument.
This is especially the case where callbacks are involved.