Feature: libcrmcommon:  Support old-style XML output.
This adds a legacy_xml option to the XML formatter.  This hidden command
line option is forced if --as-xml is given in crm_mon and results in
slightly different formatting of the output XML, especially around
lists.  This option only really makes sense in the crm_mon program as
that's the only one that previously supported this style.

Note also the introduction of mon_output_legacy_xml.  This is only there
to make command line handling simpler.  Once we know which format is
requested, this will be set to mon_output_xml for the rest of the
program.