Build: GNUmakefile: use same archive name for make dist and make export
... as was intended according to the pre-existing comment.

Avoid using the DIST_ARCHIVES make variable, which is undocumented and internal
to automake (and removed in the automake-NG fork). Instead, change the distdir
name so that make dist automatically uses the desired tarball name.

Since this changes the source name for the spec file, we have to use the
short commit hash in the spec file now.