Test: regression testing for acls-2-render.xsl (+ acls-render-cfg.xsl)
Note that insufficiencies in concurrent libxml2 regarding RelaxNG
validation were discovered (and fixed, see the in-code references),
so preferred way to run it for now (until the fixes are propagated) is:

  env RNGVALIDATOR="jing -i" ./regression.sh acls_render

(OTOH, this new test suite is not run by default, since the
respective runner doesn't intentionally start with "test" prefix;
the same underlying problem is prevented in Travis CI run context
with explicit avoidance of pacemaker-acls-*.rng files)

The point of these tests is that we likewise test the full
validity of the original file (as would be obtained with
"cibadmin --acls-render=ns-full -U ...", something that we could
actually test as well, for a full round-trip coverage).
For that to happen, we massage selected pacemaker schemas into
a single file that we subsequently transform so that all the
elements and attributes need to be namespaced with one of the
dedicated namespace-markers (reminder: writable/readable/denied).
Outcome of said effort is also checked in, since "jing" tool
may not always be available.

TODO: which files shall be marked for distribution, etc.