Feature: tools: Use formatted output for nodes and their attrs.
This is straightforward, with the exception of add_extra_info.  This
function is print_attr_msg repurposed to simply return whether or not it
found extra info that needs to be printed.  If so, the expected_score
parameter contains what is needed to do the proper printing.

Note that this would have to be significantly changed if anything
besides ping resources were to be supported.