Build: GNUmakefile: improve coverity targets
Most importantly, add the ability to specify the coverity aggressiveness level,
and put outputs in the build directory. Otherwise mostly refactoring for
best practices.

The public coverity instance changed its upload process, so we can't do a
simple curl anymore. We haven't been using it anyway, so just echo what needs
to be done for that case.