Build: rpm: option to put RPM artifacts in build tree subdirectory
Define a new variable, RPMDEST.

If RPMDEST=toplevel (the default, to preserve backward compatibility), RPM
artifacts will be put in the same locations they have been up to this point
(RPM sources, spec file, and source rpm will be in the top-level build
directory, while everything else such as the RPM build directory and location
of generated RPMs will be in the usual default locations).

If RPMDEST=subtree, RPM artifacts will be put in the "rpm" subdirectory of the
build tree. This allows "make rpm" to be compatible with VPATH builds, and will
be useful for CI.