Build: rpm: move SuSE modifications directly into spec file
Previously, building a spec for SuSE involved using sed to substitute package
names and such. Now, that is done directly in the spec by checking the presence
of %{suse_version}.

This allows GNUmakefile to be greatly simplified, and allows us to avoid
needing an intermediate pacemaker-DISTRO.spec file.

It's also much less brittle; it was too easy to forget the sed in GNUmakefile
when modifying the spec. Making the SuSE spec has actually been broken since
2.0.0 via commit d29e882a for this reason. This commit fixes that regression.