Build: add truly noarch indirection towards pacemaker schemas
This need stems from the downstream parcelling of the bits (where
it irrefutably, unlike with upstream one trying to be as suitable
and universal preimage as possible [hence this upstream change],
interacts with other software packages such as clufter [see also
6ef3f776b]), since previously, any external package only interested
in the actual schemas or transformation stylesheets would (in
systemically constructed approach not relying on any kind of
pacemaker-related hardwiring, which is recommended for obvious
reasons) query pacemaker-cib.pc for schemadir variable.  However,
for that file to be present, the avalanche of packages starting
with pacemaker devel package (and consequently corosync libraries,
libqb, ...) is needed for that, adding up to the from-ground-up
requirements unreasonably.

The solution is to devise a brand new pkg-config file
pacemaker-schemas.pc that would only serve that indirection
purpose.  It will overlap with pacemaker-cib.pc at the mentioned
item, however, this new file is the only to receive all the updates
as such top directory becomes sub-divided, as already manifested
with a brand new schemadir4api variable.

All the users of schemas/transformations alone are hence suggested
to primarily aim (arch-independent) pacemaker-schemas.pc first now,
and pacemaker-cib.pc only as a fallback when older version of
pacemaker lacking the former is encountered.

Regarding the change in the upstream spec file, note that in practice
(with some downstreams, at least, observed with Fedora), this adds
a new automatic dependency of pacemaker-schemas subpackage on pkg-config
tool, however it is rather self-contained (unlike the transitive cluster
library dependencies), and is likely to reside on real use-systems/roots
anyway.  If that'd be a blocker for some minimalistic use cases, it
shall be straightforward to turn that into something like Recommends:.