Doc: fix stop-all-resources self-documenting text
cluster option "stop-all-resources" behaviour was changed in f4d8a29d6d
at 2010. It was reverted by ecd95f7f10 at 2012, but only left the
self-documenting text behind. It makes 'man pengine' and 'man
pacemaker-schedulerd' conflict with the actual behaviour since then.
BTW, "Pacemaker Explained" is up-to-date and in the good shape for it.

Signed-off-by: Roger Zhou <zzhou@suse.com>