Maint: new file IDEAS.md to dump thoughts about future extensions
Start thematically, related to speed and integration concerns.

In a sense, it is an (fully independent and putting less accent on "do"
and more on "ideas", though) reincarnation of TODO.markdown that got
stale and eventually dropped (6b1d65bc4).  Similar story cannot be fully
excluded here, but for any history digging, it's helpful to have some
sort of wishful/future-anticipating thinking materialized for
a retro-perusal, since it allows for better understanding of the
zeigeist/mindset around the project at particular time.