Build: option to keep RPMs in build tree
This is a roll-up backport of master branch changes to .gitignore, GNUmakefile,
Makefile.am, rpmlintrc, and pacemaker.spec.in.

While normally we would not modify the 2.0 branch until ready to make a release
candidate, these changes are needed for CI purposes. The plan is to revert this
commit before pulling in the master branch for the eventual release candidate.