Low: libpe_status: check for parameter changes for any unexpired operation
Last failures of start or monitor are cleared if the resource's parameters
change. Previously, this check would be skipped when a resource has a
failure-timeout, an expired operation, and an unexpired fail count. This
appears to be an oversight, though the situation should be impossible in a
properly functioning cluster.