Fix: libstonith: Sort history before stonith__later_succeeded.
Add a comment on that function explaining the need to sort.  Then, move
sort_stonith_history into libstonith as an internal function and rename
it.  Finally, change all the callers of stonith__later_succeeded to
first sort.  Note that this also fixes a bug where the wrong type was
being passed to the stonith-event messages.