Feature: Pacemaker Remote: allow environment variable file for PID 1
Environment variables can be set in bundle containers via the extra options
configured to pass to the container launcher. However, this does not allow
environment variables whose value might differ by host (such as
PCMK_remote_address).

Now, when pacemaker-remoted runs as PID 1, it will look for a file
/etc/pacemaker/pcmk-init.env, and if it exists, parse name/value pairs from it
using a shell-like syntax, and set those in the environment. Users can mount
a file on the container host as /etc/pacemaker/pcmk-init.env inside the
container, via the bundle's storage-mapping, and the contents of the file can
vary by host.