Fix: libcrmcommon:  Be more lenient when freeing private data.
"crm_mon --output-as=xml --output-to=/nonexistent/file.xml" crashes on
exit because the finish functions try to dereference priv, which wil be
NULL for that command line.  So just return without doing anything.
Additionally, if the messages table is NULL, g_hash_table_destroy will
also crash.  Protect against that.