Fix: tools: Add blank lines in crm_mon's text output.
In order to do that, however, we need to know if the previous output
function actually printed anything.  So, change a bunch of them to
return a gboolean.  Then add tests to the beginning of each to determine
whether there's anything to output and return the appropriate value.

This also has the effect of skipping sections if they would be empty.