Build: RPM: update RHEL 8 directory for documentation package
In older RHEL versions, the main pacemaker package's %doc files would be
installed in /usr/share/doc/pacemaker-VERSION, but in RHEL 8, they are
installed in /usr/share/doc/pacemaker. This causes conflicts with identically
named files (COPYING and ChangeLog) in the pacemaker-doc subpackage, so
move the latter's destination to /usr/share/doc/pacemaker-doc for RHEL 8.