Doc: ChangeLog: improve summary
Don't include merge commits when counting the number of changesets in a
release (this has not been applied retroactively to old releases).

Also, newer versions of git diff apparently print a carriage return such that
anything printed without a newline before it ("Diff:" in our case) doesn't show
up. So, print a newline after "Diff:". The lines can be joined manually if
desired.