Fix: libcrmcommon: Wrap nested lists in <li> tags.
Previously, the generated HTML for nested lists looked something like
this:

<h2>Outer List</h2>
<ul>
<li>First Item</li>
<h2>Inner List</h2>
<ul>
<li>First Inner Item</li>
</ul>
</ul>

This isn't quite right.  The inner list is an element of the outer list,
and so should be wrapped in <li> tags.  This generates HTML like this:

<h2>Outer List</h2>
<ul>
<li>First Item</li>
<li>
<h2>Inner List</h2>
<ul>
<li>First Inner Item</li>
</ul>
</li>
</ul>

The new HTML definitely has a lot more list tags, especially when it
comes to node attributes and operations.  However, I believe this is
also more correct.